pc-defender-etl/requirements.txt
prismacloud_api
direct-redis
psycopg2-binary
redis
pandas
